action type enum changes
