+++ v2/pkg/protocols/headless/engine/page_actions.go
@@ -24,7 +24,7 @@ func (p *Page) ExecuteActions(baseURL *url.URL, actions []*Action) (map[string]s
 
 	outData := make(map[string]string)
 	for _, act := range actions {
-		actionType := ActionStringToAction[act.ActionType]
+		actionType := ActionStringToAction[act.ActionType.String()]
 
 		switch actionType {
 		case ActionNavigate:
