+++ v2/pkg/protocols/headless/engine/page_actions_test.go
@@ -28,7 +28,7 @@ func TestActionNavigate(t *testing.T) {
 		</body>
 	</html>`
 
-	actions := []*Action{{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}}, {ActionType: "waitload"}}
+	actions := []*Action{{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}}, {ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}}}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
 		require.Nil(t, err, "could not run page actions")
@@ -50,10 +50,11 @@ func TestActionScript(t *testing.T) {
 
 	t.Run("run-and-results", func(t *testing.T) {
 		actions := []*Action{
-			{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-			{ActionType: "waitload"},
-			{ActionType: "script", Name: "test", Data: map[string]string{"code": "window.test"}},
+			{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+			{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+			{ActionType: ActionTypeHolder{ActionType: ActionType(ActionScript)}, Name: "test", Data: map[string]string{"code": "window.test"}},
 		}
+
 		testHeadlessSimpleResponse(t, response, actions, timeout, func(page *Page, err error, out map[string]string) {
 			require.Nil(t, err, "could not run page actions")
 			require.Equal(t, "Nuclei Test Page", page.Page().MustInfo().Title, "could not navigate correctly")
@@ -63,10 +64,10 @@ func TestActionScript(t *testing.T) {
 
 	t.Run("hook", func(t *testing.T) {
 		actions := []*Action{
-			{ActionType: "script", Data: map[string]string{"code": "window.test = 'some-data';", "hook": "true"}},
-			{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-			{ActionType: "waitload"},
-			{ActionType: "script", Name: "test", Data: map[string]string{"code": "window.test"}},
+			{ActionType: ActionTypeHolder{ActionType: ActionType(ActionScript)}, Data: map[string]string{"code": "window.test = 'some-data';", "hook": "true"}},
+			{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+			{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+			{ActionType: ActionTypeHolder{ActionType: ActionType(ActionScript)}, Name: "test", Data: map[string]string{"code": "window.test"}},
 		}
 		testHeadlessSimpleResponse(t, response, actions, timeout, func(page *Page, err error, out map[string]string) {
 			require.Nil(t, err, "could not run page actions")
@@ -87,9 +88,9 @@ func TestActionClick(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "click", Data: map[string]string{"selector": "button"}}, // Use css selector for clicking
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionClick)}, Data: map[string]string{"selector": "button"}}, // Use css selector for clicking
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -120,9 +121,9 @@ func TestActionRightClick(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "rightclick", Data: map[string]string{"selector": "button"}}, // Use css selector for clicking
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionRightClick)}, Data: map[string]string{"selector": "button"}}, // Use css selector for clicking
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -145,9 +146,9 @@ func TestActionTextInput(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "text", Data: map[string]string{"selector": "input", "value": "test"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionTextInput)}, Data: map[string]string{"selector": "input", "value": "test"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -162,9 +163,9 @@ func TestActionTextInput(t *testing.T) {
 
 func TestActionHeadersChange(t *testing.T) {
 	actions := []*Action{
-		{ActionType: "setheader", Data: map[string]string{"part": "request", "key": "Test", "value": "Hello"}},
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionSetHeader)}, Data: map[string]string{"part": "request", "key": "Test", "value": "Hello"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
 	}
 
 	handler := func(w http.ResponseWriter, r *http.Request) {
@@ -189,9 +190,9 @@ func TestActionScreenshot(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "screenshot", Data: map[string]string{"to": "test"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionScreenshot)}, Data: map[string]string{"to": "test"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -214,9 +215,9 @@ func TestActionTimeInput(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "time", Data: map[string]string{"selector": "input", "value": "2006-01-02T15:04:05Z"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionTimeInput)}, Data: map[string]string{"selector": "input", "value": "2006-01-02T15:04:05Z"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -242,9 +243,9 @@ func TestActionSelectInput(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "select", Data: map[string]string{"by": "x", "xpath": "//select[@id='test']", "value": "Test2", "selected": "true"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionSelectInput)}, Data: map[string]string{"by": "x", "xpath": "//select[@id='test']", "value": "Test2", "selected": "true"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -265,9 +266,9 @@ func TestActionFilesInput(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "files", Data: map[string]string{"selector": "input", "value": "test1.pdf"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionFilesInput)}, Data: map[string]string{"selector": "input", "value": "test1.pdf"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -291,8 +292,8 @@ func TestActionWaitLoad(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -316,8 +317,8 @@ func TestActionGetResource(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "getresource", Data: map[string]string{"by": "x", "xpath": "//img[@id='test']"}, Name: "src"},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionGetResource)}, Data: map[string]string{"by": "x", "xpath": "//img[@id='test']"}, Name: "src"},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -336,8 +337,8 @@ func TestActionExtract(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "extract", Data: map[string]string{"by": "x", "xpath": "//button[@id='test']"}, Name: "extract"},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionExtract)}, Data: map[string]string{"by": "x", "xpath": "//button[@id='test']"}, Name: "extract"},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -355,8 +356,8 @@ func TestActionSetMethod(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "setmethod", Data: map[string]string{"part": "x", "method": "SET"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionSetMethod)}, Data: map[string]string{"part": "x", "method": "SET"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -367,9 +368,9 @@ func TestActionSetMethod(t *testing.T) {
 
 func TestActionAddHeader(t *testing.T) {
 	actions := []*Action{
-		{ActionType: "addheader", Data: map[string]string{"part": "request", "key": "Test", "value": "Hello"}},
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionAddHeader)}, Data: map[string]string{"part": "request", "key": "Test", "value": "Hello"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
 	}
 
 	handler := func(w http.ResponseWriter, r *http.Request) {
@@ -386,11 +387,11 @@ func TestActionAddHeader(t *testing.T) {
 
 func TestActionDeleteHeader(t *testing.T) {
 	actions := []*Action{
-		{ActionType: "addheader", Data: map[string]string{"part": "request", "key": "Test1", "value": "Hello"}},
-		{ActionType: "addheader", Data: map[string]string{"part": "request", "key": "Test2", "value": "World"}},
-		{ActionType: "deleteheader", Data: map[string]string{"part": "request", "key": "Test2"}},
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionAddHeader)}, Data: map[string]string{"part": "request", "key": "Test1", "value": "Hello"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionAddHeader)}, Data: map[string]string{"part": "request", "key": "Test2", "value": "World"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionDeleteHeader)}, Data: map[string]string{"part": "request", "key": "Test2"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
 	}
 
 	handler := func(w http.ResponseWriter, r *http.Request) {
@@ -407,9 +408,9 @@ func TestActionDeleteHeader(t *testing.T) {
 
 func TestActionSetBody(t *testing.T) {
 	actions := []*Action{
-		{ActionType: "setbody", Data: map[string]string{"part": "request", "body": "hello"}},
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionSetBody)}, Data: map[string]string{"part": "request", "body": "hello"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
 	}
 
 	handler := func(w http.ResponseWriter, r *http.Request) {
@@ -435,10 +436,10 @@ func TestActionKeyboard(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitload"},
-		{ActionType: "click", Data: map[string]string{"selector": "input"}},
-		{ActionType: "keyboard", Data: map[string]string{"keys": "Test2"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitLoad)}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionClick)}, Data: map[string]string{"selector": "input"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionKeyboard)}, Data: map[string]string{"keys": "Test2"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -461,8 +462,8 @@ func TestActionSleep(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "sleep", Data: map[string]string{"duration": "2"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionSleep)}, Data: map[string]string{"duration": "2"}},
 	}
 
 	testHeadlessSimpleResponse(t, response, actions, 20*time.Second, func(page *Page, err error, out map[string]string) {
@@ -484,8 +485,8 @@ func TestActionWaitVisible(t *testing.T) {
 		</html>`
 
 	actions := []*Action{
-		{ActionType: "navigate", Data: map[string]string{"url": "{{BaseURL}}"}},
-		{ActionType: "waitvisible", Data: map[string]string{"by": "x", "xpath": "//button[@id='test']"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionNavigate)}, Data: map[string]string{"url": "{{BaseURL}}"}},
+		{ActionType: ActionTypeHolder{ActionType: ActionType(ActionWaitVisible)}, Data: map[string]string{"by": "x", "xpath": "//button[@id='test']"}},
 	}
 
 	t.Run("wait for an element being visible", func(t *testing.T) {
